+++ src/main/java/by/dzmitry/test_task/service/numberplate/NumberServiceImpl.java
@@ -59,6 +59,9 @@ public class NumberServiceImpl implements NumberplateService {
         if (numberplate.isEmpty()) {
             throw new CustomException("Numberplate not found");
         }
+        if(numberplate.get().getCar() !=null){
+            throw new CustomException("Can't delete uses numberplate with");
+        }
         numberplateDao.deleteById(numberplateId);
     }
